requirements.txt
cryptography>=36.0.0
discord.py>=2.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
google-genai>=0.4.0
anthropic>=0.20.0
openai>=1.12.0  # Updated from 0.27.8 to support new API pattern
pytz>=2023.3
tzlocal>=5.0.1
json5>=0.9.14
aiohttp>=3.9.0
requests>=2.31.0
sqlite3-api>=2.0  # Renamed from SQLite3
Pillow>=10.0.0
python-dateutil>=2.8.2
asyncio>=3.4.3
psutil>=5.9.5
PyYAML>=6.0.1
# Dashboard requirements
flask[async]>=2.3.0 # Added async extra
flask-cors>=4.0.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.1.0
# Discord SSO requirements
flask-login>=0.6.2
oauthlib>=3.2.2
requests-oauthlib>=1.3.1
flask-session>=0.5.0
itsdangerous>=2.1.2
# Additional dashboard dependencies
numpy>=1.26.0
werkzeug>=2.3.0
gunicorn>=21.2.0  # For production deployment
waitress>=2.1.2   # WSGI server for production